apply filter from authorization in sync
that's to sync only user workbench notes
